refactor: reorganized files to define a "core"
I'd like to have a folder where all the OAuth protocol logic is included
"ClientServices" would orchestrate using the core
