+++ projects/angular-simple-oidc/src/lib/token-endpoint-client.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { switchMap, take, map } from 'rxjs/operators';
-import { TokenRequestResult, DecodedIdentityToken } from './models';
-import { TokenValidationService } from './token-validation.service';
-import { OidcDiscoveryDocClient } from '../discovery-document/oidc-discovery-doc-client.service';
-import { TokenHelperService } from './token-helper.service';
-import { ValidationResult } from './validation-result';
+import { TokenRequestResult, DecodedIdentityToken } from './token/models';
+import { TokenValidationService } from './token/token-validation.service';
+import { OidcDiscoveryDocClient } from './discovery-document/oidc-discovery-doc-client.service';
+import { TokenHelperService } from './token/token-helper.service';
+import { ValidationResult } from './token/validation-result';
 
 interface TokenEndpointResponse {
     access_token?: string;
