+++ projects/angular-simple-oidc/src/lib/token-storage.service.ts
@@ -2,9 +2,9 @@ import { Injectable, Inject } from '@angular/core';
 import {
     TokenStorageKeys, LocalState,
     TokenRequestResult, DecodedIdentityToken
-} from './models';
+} from './token/models';
 import { of, BehaviorSubject } from 'rxjs';
-import { LOCAL_STORAGE_REF } from '../constants';
+import { LOCAL_STORAGE_REF } from './constants';
 
 // @dynamic
 @Injectable()
