+++ projects/angular-simple-oidc/src/lib/token/angular-simple-oidc-core.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { TokenCryptoService } from './token-crypto.service';
+import { TokenHelperService } from './token-helper.service';
+import { TokenValidationService } from './token-validation.service';
+import { TokenUrlService } from './token-url.service';
+import { RefreshTokenValidationService } from 'angular-simple-oidc';
+
+@NgModule({
+  imports: [
+  ],
+  providers: [
+    TokenCryptoService,
+    TokenUrlService,
+    TokenHelperService,
+    TokenValidationService,
+    RefreshTokenValidationService,
+  ],
+  declarations: [],
+})
+export class AngularSimpleOidcCoreModule { }
