+++ projects/angular-simple-oidc/src/public-api.ts
@@ -5,22 +5,24 @@
 export { AngularSimpleOidcModule } from './lib/angular-simple-oidc.module';
 export { AuthGuard } from './lib/guards/auth.guard';
 export { AuthService } from './lib/auth.service';
+
+export { TokenStorageService } from './lib/token-storage.service';
 export { OidcDiscoveryDocClient } from './lib/discovery-document/oidc-discovery-doc-client.service';
 
 export { TokenCryptoService } from './lib/token/token-crypto.service';
 export { TokenHelperService } from './lib/token/token-helper.service';
 export { TokenUrlService } from './lib/token/token-url.service';
-export { TokenStorageService } from './lib/token/token-storage.service';
 export { TokenValidationService } from './lib/token/token-validation.service';
 
 export { AuthConfig } from './lib/config/models';
-export { DiscoveryDocument, JWTKey, JWTKeys } from './lib/discovery-document/models';
 export {
+    DiscoveryDocument,
+    JWTKey,
+    JWTKeys,
     LocalState,
     DecodedIdentityToken,
     TokenRequestResult
 } from './lib/token/models';
 
-export { RefreshTokenModule } from './lib/refresh-token/refresh-token.module';
-export { RefreshTokenClient } from './lib/refresh-token/refresh-token-client.service';
-export { RefreshTokenValidationService } from './lib/refresh-token/refresh-token-validation.service';
+export { RefreshTokenClient } from './lib/refresh-token-client.service';
+export { RefreshTokenValidationService } from './lib/token/refresh-token/refresh-token-validation.service';
